# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de trading IA para IQ Option
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface MessageResponse {
  message: string;
  success: boolean;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface LoginResponse {
  success: boolean;
  message: string;
  accountType?: string;
  balance?: number;
}

export interface AuthStatus {
  connected: boolean;
  email?: string;
  accountType?: string;
  balance?: number;
  usingRealData?: boolean;
}

export interface Balance {
  balance: number;
  currency: string;
  accountType: string;
}

export type SwitchAccountRequestAccountType =
  (typeof SwitchAccountRequestAccountType)[keyof typeof SwitchAccountRequestAccountType];

export const SwitchAccountRequestAccountType = {
  REAL: "REAL",
  PRACTICE: "PRACTICE",
} as const;

export interface SwitchAccountRequest {
  accountType: SwitchAccountRequestAccountType;
}

export interface Candle {
  time: number;
  open: number;
  close: number;
  high: number;
  low: number;
  volume?: number;
}

export interface CandlesResponse {
  asset: string;
  timeframe: number;
  candles: Candle[];
}

export interface AssetCategory {
  name: string;
  assets: string[];
}

export interface AssetsResponse {
  categories: AssetCategory[];
  otcAssets: string[];
}

export interface IndicatorValues {
  rsi?: number;
  macd?: number;
  macdSignal?: number;
  macdHistogram?: number;
  bbUpper?: number;
  bbMiddle?: number;
  bbLower?: number;
  ma5?: number;
  ma10?: number;
  ma20?: number;
}

export type SignalMode = (typeof SignalMode)[keyof typeof SignalMode];

export const SignalMode = {
  OTC: "OTC",
  REAL: "REAL",
} as const;

export type SignalDirectionPA =
  (typeof SignalDirectionPA)[keyof typeof SignalDirectionPA];

export const SignalDirectionPA = {
  CALL: "CALL",
  PUT: "PUT",
  NEUTRO: "NEUTRO",
} as const;

export type SignalDirectionML =
  (typeof SignalDirectionML)[keyof typeof SignalDirectionML];

export const SignalDirectionML = {
  CALL: "CALL",
  PUT: "PUT",
  NEUTRO: "NEUTRO",
} as const;

export type SignalDirectionRSI =
  (typeof SignalDirectionRSI)[keyof typeof SignalDirectionRSI];

export const SignalDirectionRSI = {
  CALL: "CALL",
  PUT: "PUT",
  NEUTRO: "NEUTRO",
} as const;

export type SignalDirectionFinal =
  (typeof SignalDirectionFinal)[keyof typeof SignalDirectionFinal];

export const SignalDirectionFinal = {
  CALL: "CALL",
  PUT: "PUT",
  NEUTRO: "NEUTRO",
} as const;

export interface Signal {
  asset: string;
  price: number;
  mode: SignalMode;
  directionPA: SignalDirectionPA;
  confidencePA: number;
  directionML: SignalDirectionML;
  confidenceML: number;
  directionRSI: SignalDirectionRSI;
  confidenceRSI: number;
  directionFinal: SignalDirectionFinal;
  confidenceFinal: number;
  justification: string;
  indicators?: IndicatorValues;
  timestamp: string;
  category: string;
  currentPrice?: number;
  usingRealData?: boolean;
}

export interface SignalsResponse {
  signals: Signal[];
  totalAnalyzed: number;
  lastUpdate: string;
}

export interface SignalHistoryResponse {
  history: Signal[];
  total: number;
}

export type GetCandlesParams = {
  timeframe?: number;
  count?: number;
};

export type GetSignalsParams = {
  category?: string;
};