# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API de trading IA para IQ Option
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login IQ Option
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login realizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LoginResponse"
        "401":
          description: Credenciais inválidas
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /assets/{asset}/candles:
    get:
      operationId: getCandles
      tags: [assets]
      summary: Candles de um ativo
      parameters:
        - name: asset
          in: path
          required: true
          schema:
            type: string
        - name: timeframe
          in: query
          schema:
            type: integer
            default: 60
        - name: count
          in: query
          schema:
            type: integer
            default: 50
      responses:
        "200":
          description: Candles do ativo
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CandlesResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
        password:
          type: string
      required:
        - email
        - password
    LoginResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        accountType:
          type: string
        balance:
          type: number
      required:
        - success
        - message
...
    Candle:
      type: object
      properties:
        time:
          type: integer
        open:
          type: number
        close:
          type: number
        high:
          type: number
        low:
          type: number
        volume:
          type: number
      required:
        - time
        - open
        - close
        - high
        - low
    CandlesResponse:
      type: object
      properties:
        asset:
          type: string
        timeframe:
          type: integer
        candles:
          type: array
          items:
            $ref: "#/components/schemas/Candle"
      required:
        - asset
        - timeframe
        - candles
...