# lib/api-zod/src/generated/api.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API de trading IA para IQ Option
 * OpenAPI spec version: 0.1.0
 */
import * as zod from "zod";

/**
 * Returns server health status
 * @summary Health check
 */
export const HealthCheckResponse = zod.object({
  status: zod.string(),
});

/**
 * @summary Login IQ Option
 */
export const LoginBody = zod.object({
  email: zod.string(),
  password: zod.string(),
});

export const LoginResponse = zod.object({
  success: zod.boolean(),
  message: zod.string(),
  accountType: zod.string().optional(),
  balance: zod.number().optional(),
});

/**
 * @summary Status de conexão
 */
export const GetAuthStatusResponse = zod.object({
  connected: zod.boolean(),
  email: zod.string().optional(),
  accountType: zod.string().optional(),
  balance: zod.number().optional(),
  usingRealData: zod.boolean().optional(),
});

/**
 * @summary Logout IQ Option
 */
export const LogoutResponse = zod.object({
  message: zod.string(),
  success: zod.boolean(),
});

/**
 * @summary Saldo da conta
 */
export const GetBalanceResponse = zod.object({
  balance: zod.number(),
  currency: zod.string(),
  accountType: zod.string(),
});

/**
 * @summary Trocar tipo de conta (REAL/PRACTICE)
 */
export const SwitchAccountTypeBody = zod.object({
  accountType: zod.enum(["REAL", "PRACTICE"]),
});

export const SwitchAccountTypeResponse = zod.object({
  message: zod.string(),
  success: zod.boolean(),
});

/**
 * @summary Lista de ativos disponíveis
 */
export const GetAssetsResponse = zod.object({
  categories: zod.array(
    zod.object({
      name: zod.string(),
      assets: zod.array(zod.string()),
    }),
  ),
  otcAssets: zod.array(zod.string()),
});

/**
 * @summary Candles de um ativo
 */
export const GetCandlesParams = zod.object({
  asset: zod.coerce.string(),
});

export const getCandlesQueryTimeframeDefault = 60;
export const getCandlesQueryCountDefault = 50;

export const GetCandlesQueryParams = zod.object({
  timeframe: zod.coerce.number().default(getCandlesQueryTimeframeDefault),
  count: zod.coerce.number().default(getCandlesQueryCountDefault),
});

export const GetCandlesResponse = zod.object({
  asset: zod.string(),
  timeframe: zod.number(),
  candles: zod.array(
    zod.object({
      time: zod.number(),
      open: zod.number(),
      close: zod.number(),
      high: zod.number(),
      low: zod.number(),
      volume: zod.number().optional(),
    }),
  ),
});

/**
 * @summary Sinais de todos os ativos monitorados
 */
export const GetSignalsQueryParams = zod.object({
  category: zod.coerce.string().optional(),
});

export const GetSignalsResponse = zod.object({
  signals: zod.array(
    zod.object({
      asset: zod.string(),
      price: zod.number(),
      mode: zod.enum(["OTC", "REAL"]),
      directionPA: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidencePA: zod.number(),
      directionML: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidenceML: zod.number(),
      directionRSI: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidenceRSI: zod.number(),
      directionFinal: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidenceFinal: zod.number(),
      justification: zod.string(),
      indicators: zod
        .object({
          rsi: zod.number().optional(),
          macd: zod.number().optional(),
          macdSignal: zod.number().optional(),
          macdHistogram: zod.number().optional(),
          bbUpper: zod.number().optional(),
          bbMiddle: zod.number().optional(),
          bbLower: zod.number().optional(),
          ma5: zod.number().optional(),
          ma10: zod.number().optional(),
          ma20: zod.number().optional(),
        })
        .optional(),
      timestamp: zod.string(),
      category: zod.string(),
      currentPrice: zod.number().optional(),
      usingRealData: zod.boolean().optional(),
    }),
  ),
  totalAnalyzed: zod.number(),
  lastUpdate: zod.string(),
});

/**
 * @summary Sinal de um ativo específico
 */
export const GetSignalByAssetParams = zod.object({
  asset: zod.coerce.string(),
});

export const GetSignalByAssetResponse = zod.object({
  asset: zod.string(),
  price: zod.number(),
  mode: zod.enum(["OTC", "REAL"]),
  directionPA: zod.enum(["CALL", "PUT", "NEUTRO"]),
  confidencePA: zod.number(),
  directionML: zod.enum(["CALL", "PUT", "NEUTRO"]),
  confidenceML: zod.number(),
  directionRSI: zod.enum(["CALL", "PUT", "NEUTRO"]),
  confidenceRSI: zod.number(),
  directionFinal: zod.enum(["CALL", "PUT", "NEUTRO"]),
  confidenceFinal: zod.number(),
  justification: zod.string(),
  indicators: zod
    .object({
      rsi: zod.number().optional(),
      macd: zod.number().optional(),
      macdSignal: zod.number().optional(),
      macdHistogram: zod.number().optional(),
      bbUpper: zod.number().optional(),
      bbMiddle: zod.number().optional(),
      bbLower: zod.number().optional(),
      ma5: zod.number().optional(),
      ma10: zod.number().optional(),
      ma20: zod.number().optional(),
    })
    .optional(),
  timestamp: zod.string(),
  category: zod.string(),
  currentPrice: zod.number().optional(),
  usingRealData: zod.boolean().optional(),
});

/**
 * @summary Histórico de sinais gerados
 */
export const GetSignalHistoryResponse = zod.object({
  history: zod.array(
    zod.object({
      asset: zod.string(),
      price: zod.number(),
      mode: zod.enum(["OTC", "REAL"]),
      directionPA: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidencePA: zod.number(),
      directionML: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidenceML: zod.number(),
      directionRSI: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidenceRSI: zod.number(),
      directionFinal: zod.enum(["CALL", "PUT", "NEUTRO"]),
      confidenceFinal: zod.number(),
      justification: zod.string(),
      indicators: zod
        .object({
          rsi: zod.number().optional(),
          macd: zod.number().optional(),
          macdSignal: zod.number().optional(),
          macdHistogram: zod.number().optional(),
          bbUpper: zod.number().optional(),
          bbMiddle: zod.number().optional(),
          bbLower: zod.number().optional(),
          ma5: zod.number().optional(),
          ma10: zod.number().optional(),
          ma20: zod.number().optional(),
        })
        .optional(),
      timestamp: zod.string(),
      category: zod.string(),
      currentPrice: zod.number().optional(),
      usingRealData: zod.boolean().optional(),
    }),
  ),
  total: zod.number(),
});